+++ src/main/java/de/samply/form/template/FormKey.java
@@ -6,7 +6,8 @@ import lombok.Getter;
 public enum FormKey {
 
     FIELDS("fields"),
-    DATA_TYPE_CLASS("DataType");
+    DATA_TYPE_CLASS("DataType"),
+    CURRENT_DATE("currentDate");
 
     private final String text;